requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Vector Database & Embeddings - FIXED VERSIONS
chromadb==0.4.22
sentence-transformers==2.2.2
huggingface-hub==0.19.4
transformers==4.35.2

# Message Queue
redis==5.0.1

# HTTP Client
httpx==0.24.1

# AI/LLM Framework
langgraph
langchain

# Data Processing
numpy==1.24.3
pandas==2.1.0

# Visualization
plotly==5.18.0

# Authentication
PyJWT==2.10.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code Quality
black==23.12.0
flake8==6.1.0
pre-commit==3.6.0

# # RPC
# python-jsonrpcy
